+++ actions/show.py
@@ -17,8 +17,5 @@ class Show(action.TerraformBaseAction):
         os.chdir(plan_path)
         self.terraform.terraform_bin_path = terraform_exec
         return_code, stdout, stderr = self.terraform.show('-no-color')
-        output = stdout + "\n" + stderr
-        if return_code == 0:
-            return (True, output)
-        else:
-            return (False, output)
+
+        return self.check_result(return_code, stdout, stderr)